fix: rename capus.py to avoid module name conflict
