+++ campus/client/__init__.py
@@ -5,7 +5,7 @@ Campus Client Package
 Provides unified Campus client interface.
 """
 
-from campus.client.campus import Campus
+from campus.client.core import Campus
 from campus.client.errors import (
     CampusClientError,
     AuthenticationError,
